ai_engine/requirements.txt
fastapi
uvicorn
pydantic
langchain
pymupdf
python-multipart
google-generativeai